Fix lint bug + Date Value

+++ synthesized/common/values/date.py
@@ -158,10 +158,10 @@ class DateValue(ContinuousValue):
         assert len(xs) == 5
         self.build()
         xs[0] = super().unify_inputs(xs=xs[0: 1])
-        xs[1] = self.hour.unify_inputs(xs=xs[1: 2])
-        xs[2] = self.dow.unify_inputs(xs=xs[2: 3])
-        xs[3] = self.day.unify_inputs(xs=xs[3: 4])
-        xs[4] = self.month.unify_inputs(xs=xs[4: 5])
+        xs[1] = self.hour.unify_inputs(xs=tf.cast(xs[1: 2], dtype=tf.int64))
+        xs[2] = self.dow.unify_inputs(xs=tf.cast(xs[2: 3], dtype=tf.int64))
+        xs[3] = self.day.unify_inputs(xs=tf.cast(xs[3: 4], dtype=tf.int64))
+        xs[4] = self.month.unify_inputs(xs=tf.cast(xs[4: 5], dtype=tf.int64))
         return tf.concat(values=xs, axis=1)
 
     # TODO: skip last and assume absolute value
